add gui

+++ src/file/Cache.cpp
@@ -2,6 +2,10 @@
 #include "Exceptions.h"
 #include "Store.h"
 
+Cache::Cache() : m_cache(), m_store(nullptr)
+{
+}
+
 Cache::Cache(Store *store) : m_cache(), m_store(store)
 {
     store->open();
@@ -14,14 +18,16 @@ Cache::Cache(Store *store) : m_cache(), m_store(store)
 Cache::~Cache()
 {
     save();
-    delete m_store;
+    if (m_store != nullptr) {
+        delete m_store;
+    }
 }
 
 const std::string &Cache::get(const std::string &name)
 {
     try {
         auto &section = m_cache.at(name);
-        if (section.readDirty) {
+        if (m_store != nullptr && section.readDirty) {
             m_store->readSection(name, section.content);
         }
         section.readDirty = false;
@@ -41,13 +47,17 @@ void Cache::put(const std::string &name, const std::string &content)
 
 void Cache::remove(const std::string &name)
 {
-    m_store->deleteSection(name);
+    if (m_store != nullptr) {
+        m_store->deleteSection(name);
+    }
     m_cache.erase(name);
 }
 
 void Cache::removePrefix(const std::string &prefix)
 {
-    m_store->deleteSectionPrefix(prefix);
+    if (m_store != nullptr) {
+        m_store->deleteSectionPrefix(prefix);
+    }
     for (auto &[name, section] : m_cache) {
         if (name.starts_with(prefix)) {
             m_cache.erase(name);
@@ -57,12 +67,17 @@ void Cache::removePrefix(const std::string &prefix)
 
 void Cache::removeAll()
 {
-    m_store->deleteAllSections();
+    if (m_store != nullptr) {
+        m_store->deleteAllSections();
+    }
     m_cache.clear();
 }
 
 void Cache::save()
 {
+    if (m_store == nullptr) {
+        return;
+    }
     for (auto &[name, section] : m_cache) {
         if (section.writeDirty) {
             m_store->writeSection(name, section.content);
@@ -74,17 +89,24 @@ void Cache::save()
 
 void Cache::saveAs(Store *store)
 {
-    if (*m_store != *store) {
+    if (store == nullptr) {
+        return;
+    }
+    if (m_store == nullptr || *m_store != *store) {
         for (auto &[name, section] : m_cache) {
             if (section.readDirty) {
                 m_store->readSection(name, section.content);
             }
         }
-        delete m_store;
-        store->create();
-        m_store = store;
+        try {
+            store->create();
+            delete m_store;
+            m_store = store;
+        } catch (FileOpen &e) {
+            throw e;
+        }
+        save();
     }
-    save();
 }
 
 void Cache::forEach(std::function<bool(const std::string &)> callback) const
